Rename Getter to View to more aptly represent its purpose
