+++ src/main/java/net/kyori/registry/BiRegistry.java
@@ -31,7 +31,7 @@ import org.checkerframework.checker.nullness.qual.NonNull;
  * @param <K> the key type
  * @param <V> the value type
  */
-public interface BiRegistry<K, V> extends BiRegistryGetter<K, V>, Registry<K, V> {
+public interface BiRegistry<K, V> extends BiRegistryView<K, V>, Registry<K, V> {
   /**
    * Creates a new bidirectional registry.
    *
