+++ src/main/java/net/kyori/registry/DefaultedBiRegistry.java
@@ -31,7 +31,7 @@ import org.checkerframework.checker.nullness.qual.NonNull;
  * @param <K> the key type
  * @param <V> the value type
  */
-public interface DefaultedBiRegistry<K, V> extends BiRegistry<K, V>, DefaultedBiRegistryGetter<K, V>, DefaultedRegistry<K, V> {
+public interface DefaultedBiRegistry<K, V> extends BiRegistry<K, V>, DefaultedBiRegistryView<K, V>, DefaultedRegistry<K, V> {
   /**
    * Creates a new bidirectional registry with a default key.
    *
