+++ src/main/java/net/kyori/registry/RegistryView.java
@@ -21,34 +21,53 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
  * SOFTWARE.
  */
-package net.kyori.registry.id;
+package net.kyori.registry;
 
-import net.kyori.registry.RegistryGetter;
 import org.checkerframework.checker.nullness.qual.NonNull;
 import org.checkerframework.checker.nullness.qual.Nullable;
 
-import java.util.OptionalInt;
+import java.util.Iterator;
+import java.util.Set;
+import java.util.stream.Stream;
+import java.util.stream.StreamSupport;
 
 /**
- * A read-only view of an id registry.
+ * A read-only view of a registry.
  *
  * @param <K> the key type
  * @param <V> the value type
+ * @see Registry
  */
-public interface IdRegistryGetter<K, V> extends RegistryGetter<K, V> {
+public interface RegistryView<K, V> extends Iterable<V> {
   /**
-   * Gets the id for {@code value}.
+   * Gets the value for {@code key}.
    *
-   * @param value the value
-   * @return the id
-   */
-  @NonNull OptionalInt id(final @NonNull V value);
-
-  /**
-   * Gets the value for {@code id}.
-   *
-   * @param id the id
+   * @param key the key
    * @return the value
    */
-  @Nullable V byId(final int id);
+  @Nullable V get(final @NonNull K key);
+
+  /**
+   * Gets a set of the keys contained in this registry.
+   *
+   * @return a set of the keys contained in this registry
+   */
+  @NonNull Set<K> keySet();
+
+  /**
+   * Creates an unmodifiable iterator of values.
+   *
+   * @return an unmodifiable iterator
+   */
+  @Override
+  @NonNull Iterator<V> iterator();
+
+  /**
+   * Creates a stream of values.
+   *
+   * @return a stream
+   */
+  default @NonNull Stream<V> stream() {
+    return StreamSupport.stream(this.spliterator(), false);
+  }
 }
