+++ src/main/java/net/kyori/registry/id/DefaultedBiIdRegistryView.java
@@ -23,21 +23,11 @@
  */
 package net.kyori.registry.id;
 
-import net.kyori.registry.DefaultedRegistryGetter;
-import org.checkerframework.checker.nullness.qual.NonNull;
-
 /**
- * A read-only view of an id registry with a default key and value.
+ * A read-only view of a bidirectional id registry with a default key and value.
  *
  * @param <K> the key type
  * @param <V> the value type
  */
-public interface DefaultedIdRegistryGetter<K, V> extends DefaultedRegistryGetter<K, V>, IdRegistryGetter<K, V> {
-  /**
-   * Gets the id for {@code value}.
-   *
-   * @param value the value
-   * @return the id
-   */
-  int idOrDefault(final @NonNull V value);
+public interface DefaultedBiIdRegistryView<K, V> extends BiIdRegistryView<K, V>, DefaultedIdRegistryView<K, V> {
 }
