+++ src/main/java/net/kyori/registry/id/IdRegistryView.java
@@ -23,11 +23,32 @@
  */
 package net.kyori.registry.id;
 
+import net.kyori.registry.RegistryView;
+import org.checkerframework.checker.nullness.qual.NonNull;
+import org.checkerframework.checker.nullness.qual.Nullable;
+
+import java.util.OptionalInt;
+
 /**
- * A read-only view of a bidirectional id registry with a default key and value.
+ * A read-only view of an id registry.
  *
  * @param <K> the key type
  * @param <V> the value type
  */
-public interface DefaultedBiIdRegistryGetter<K, V> extends BiIdRegistryGetter<K, V>, DefaultedIdRegistryGetter<K, V> {
+public interface IdRegistryView<K, V> extends RegistryView<K, V> {
+  /**
+   * Gets the id for {@code value}.
+   *
+   * @param value the value
+   * @return the id
+   */
+  @NonNull OptionalInt id(final @NonNull V value);
+
+  /**
+   * Gets the value for {@code id}.
+   *
+   * @param id the id
+   * @return the value
+   */
+  @Nullable V byId(final int id);
 }
